update：配置TS环境，完成第一份题解
