+++ README.md
@@ -1,9 +1,20 @@
-# fe-agorithm-interview
-前端算法面试常考题目解析。
+<h1>前端算法面试高频题库&题目解析</h1>
 
-一站式解决前端面试遇到的算法问题
+<h4>一站式解决前端面试遇到的算法问题，咱们只专注于前端高频算法。</h4>
 
-解决痛点：
-（1）LeetCode 题目太多，不知从何刷起？
-（2）面试题目有些是前端常考，有些是后端常考？想把时间花在刀刃上怎么办？
-（3）其他仓库为了吸引关注，会尽量把所有题目列的很’全‘（刷不完那一种），根本没有照顾到求职心切的人的心情？
+<p>解决痛点：
+<br>
+（1）LeetCode 题目太多，不知从何刷起？面试题目有些是前端常考，有些是后端常考？
+<br>
+答：本站参照 CodeTop 上的前端算法题目考察频度，抽取前100道题，知道你时间宝贵，咱们要把时间花在刀刃上。
+<br>
+（2）leetcode 上面的题解鱼龙混杂，有些直接题目直接贴代码，没有任何分析过程，看不懂？
+<br>
+答：本站会带你理解题目，并且给出解题思路，分析时间&空间复杂度，一题多解，带你彻底搞懂。
+<br>
+（3）github 有些仓库为了吸引关注数目，会尽量把所有题目列的很’全‘（刷不完那一种），根本没有照顾到求职心切的人的心情？
+<br>
+答：本站只做专精，不搞大而全。根据考试频率划分层次，让你做题时心里有数，减少求职者刷算法时的心理负担。
+<br>
+使用方法：待补充
+</p>
