--- src/03-dachang/01-reverse-link-list.js
@@ -1,4 +1,0 @@
-export default function func() {
-    console.log('配置JS环境!!!')
-    
-}
