+++ src/index.ts
@@ -1,0 +1,9 @@
+import './index.css'
+// import func from './01-nowcoder-top-101/01-reverse-link-list.js'
+
+// if (module.hot) {
+//     module.hot.accept('./01-nowcoder-top-101/01-reverse-link-list', function() {
+//         func()
+//     })
+// }
+console.log('Hello Webpack !!!')
